Added explicit version control of reactor-core
This commit removes the management of the reactor-core dependency from
Boot and delegates it to Spring Cloud Data Flow, preventing Boot from
overriding the version specified by the cloudfoundry client.  This being
said, it is now on the SCDF team to make sure the versions are in sync
(between what SCDF requires and what the cloudfoundry client requires).